refactor: changes the SurveyList to use the new custom 'useErrorHandler' hook instead of catching errors inline.
The 'useErrorHandler' hook was added to handle global error handling in
a consistent and reusable way across the application, thus avoiding
repetition and improving code maintainability. The refactoring in the
SurveyList component is necessary to adapt it to use this new hook,
improving its readability and consistency with the rest of the
application.